+++ ExpanderPi/demos/demo_adcread.py
@@ -17,6 +17,8 @@ from __future__ import absolute_import, division, print_function, \
                                                     unicode_literals
 
 import time
+import os
+import sys
 
 try:
     import ExpanderPi
@@ -40,12 +42,36 @@ def main():
 
     # set the reference voltage.  this should be set to the exact voltage
     # measured on the Expander Pi Vref pin.
-    adc.set_adc_refvoltage(4.096)
+    adc.set_adc_refvoltage(1.8)
 
+    # clear the console
+    os.system('clear')
     while True:
-        # read the voltage from channel 1 in single ended mode and print
-        print(adc.read_adc_voltage(1, 0))
-        time.sleep(0.5)
+        
+
+         # read the voltage from channels 1 to 8 single ended mode and print
+        sys.stdout.write('Pin 1:  %f   \n' % adc.read_adc_voltage(1, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 2:  %f   \n' % adc.read_adc_voltage(2, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 3:  %f   \n' % adc.read_adc_voltage(3, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 4:  %f   \n' % adc.read_adc_voltage(4, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 5:  %f   \n' % adc.read_adc_voltage(5, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 6:  %f   \n' % adc.read_adc_voltage(6, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 7:  %f   \n' % adc.read_adc_voltage(7, 0))
+        sys.stdout.flush()
+        sys.stdout.write('Pin 8:  %f   \n' % adc.read_adc_voltage(8, 0))
+        sys.stdout.flush()
+        sys.stdout.write('\033[A\033[A\033[A\033[A\033[A\033[A\033[A\033[A')
+        sys.stdout.flush()
+        
+
+        time.sleep(0.1)
+        
 
 if __name__ == "__main__":
     main()